Created separate file for covid test data
